Fifa11
listagem de usuarios, alteraçoes e exclusoes funcionando

+++ src/java/br/edu/vianna/fifa/model/domain/Campeonato.java
@@ -36,7 +36,7 @@ public class Campeonato implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(nullable = false)
-    private Integer id;
+    private Long id;
     @JoinColumn(name = "idTime", referencedColumnName = "id", nullable = false)
     @ManyToOne(optional = false)
     private Time idTime;
@@ -46,15 +46,15 @@ public class Campeonato implements Serializable {
     public Campeonato() {
     }
 
-    public Campeonato(Integer id) {
+    public Campeonato(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
