+++ src/java/br/edu/vianna/fifa/model/domain/Partida.java
@@ -40,7 +40,7 @@ public class Partida implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(nullable = false)
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @Column(nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
@@ -60,20 +60,20 @@ public class Partida implements Serializable {
     public Partida() {
     }
 
-    public Partida(Integer id) {
+    public Partida(Long id) {
         this.id = id;
     }
 
-    public Partida(Integer id, Date localTime) {
+    public Partida(Long id, Date localTime) {
         this.id = id;
         this.localTime = localTime;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
